requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary==2.9.9
gunicorn==20.1.0
django-environ==0.10.0
python-dotenv
whitenoise
openai
whisper
moviepy